pyproject.toml
[project]
name = "doeksu"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "dotenv",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langsmith>=0.0.60",
    "langchain-community",
    "langchain-text-splitters",
    "openai>=1.0.0",
    "langchain-google-genai>=0.0.5",
    "python-dotenv>=1.0.0",
    "langchain-anthropic>=0.1.0",
    "requests>=2.31.0",
    "langgraph-cli",
    "langgraph-api",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google_genai",
    "langchain-ollama",
    "pydantic>=2.0.0",
    "aiohttp",
    
    "feedparser",
    "lxml_html_clean",
    "googlenewsdecoder",
    "newspaper3k",

    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black" 

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
]
